Add spawn diagnostics for debugging monster spawning mechanics
