+++ src/managers/spawn-diagnostics.ts
@@ -147,41 +147,3 @@ export class SpawnDiagnostics {
     console.groupEnd();
   }
 }
-
-// Export a global diagnostic function for easy console access
-(window as any).spawnDiagnostics = {
-  dump: () => {
-    const gameManager = (window as any).gameManager;
-    if (gameManager?.monsterSpawnManager && gameManager?.monsterRespawnManager) {
-      const diagnostics = new SpawnDiagnostics(
-        gameManager.monsterSpawnManager,
-        gameManager.monsterRespawnManager
-      );
-      diagnostics.dumpFullDiagnostics();
-    } else {
-      console.error("Game managers not found. Is the game running?");
-    }
-  },
-  start: (intervalMs: number = 1000) => {
-    const gameManager = (window as any).gameManager;
-    if (gameManager?.monsterSpawnManager && gameManager?.monsterRespawnManager) {
-      const diagnostics = new SpawnDiagnostics(
-        gameManager.monsterSpawnManager,
-        gameManager.monsterRespawnManager
-      );
-      (window as any).spawnDiagnosticsInstance = diagnostics;
-      diagnostics.startDiagnostics(intervalMs);
-    } else {
-      console.error("Game managers not found. Is the game running?");
-    }
-  },
-  stop: () => {
-    const instance = (window as any).spawnDiagnosticsInstance;
-    if (instance) {
-      instance.stopDiagnostics();
-      delete (window as any).spawnDiagnosticsInstance;
-    } else {
-      console.error("No diagnostic instance running");
-    }
-  }
-};
